#544 Move symbol tests into the new unit-test files
