+++ src/psyclone/tests/psyGen_test.py
@@ -57,7 +57,7 @@ from psyclone.psyGen import TransInfo, Transformation, PSyFactory, NameSpace, \
     Assignment, Reference, BinaryOperation, Array, Literal, Node, IfBlock, \
     KernelSchedule, Schedule, UnaryOperation, NaryOperation, Return, \
     ACCEnterDataDirective, ACCKernelsDirective, Container
-from psyclone.psyir.symbols import DataSymbol, ContainerSymbol, SymbolTable
+from psyclone.psyir.symbols import DataSymbol, SymbolTable
 from psyclone.psyGen import GenerationError, FieldNotFoundError, \
      InternalError, HaloExchange, Invoke, DataAccess
 from psyclone.psyGen import Kern, Arguments, CodedKern
@@ -3145,7 +3145,6 @@ def test_assignment_semantic_navigation():
 
 # Test Reference class
 
-
 def test_reference_view(capsys):
     ''' Check the view and colored_text methods of the Reference class.'''
     from psyclone.psyGen import colored, SCHEDULE_COLOUR_MAP
@@ -3520,697 +3519,6 @@ def test_kernelschedule_name_setter():
     assert kschedule.name == "kname"
     kschedule.name = "newname"
     assert kschedule.name == "newname"
-
-
-# Test DataSymbol Class
-def test_symbol_initialisation():
-    '''Test that a DataSymbol instance can be created when valid arguments are
-    given, otherwise raise relevant exceptions.'''
-
-    # Test with valid arguments
-    assert isinstance(DataSymbol('a', 'real'), DataSymbol)
-    # real constants are not currently supported
-    assert isinstance(DataSymbol('a', 'integer'), DataSymbol)
-    assert isinstance(DataSymbol('a', 'integer', constant_value=0), DataSymbol)
-    assert isinstance(DataSymbol('a', 'character'), DataSymbol)
-    assert isinstance(DataSymbol('a', 'character', constant_value="hello"),
-                      DataSymbol)
-    assert isinstance(DataSymbol('a', 'boolean'), DataSymbol)
-    assert isinstance(DataSymbol('a', 'boolean', constant_value=False),
-                      DataSymbol)
-    assert isinstance(DataSymbol('a', 'real', [None]), DataSymbol)
-    assert isinstance(DataSymbol('a', 'real', [3]), DataSymbol)
-    assert isinstance(DataSymbol('a', 'real', [3, None]), DataSymbol)
-    assert isinstance(DataSymbol('a', 'real', []), DataSymbol)
-    assert isinstance(DataSymbol('a', 'real', [],
-                                 interface=DataSymbol.Argument()),
-                      DataSymbol)
-    assert isinstance(
-        DataSymbol('a', 'real', [],
-                   interface=DataSymbol.Argument(
-                       access=DataSymbol.Access.READWRITE)),
-        DataSymbol)
-    assert isinstance(
-        DataSymbol('a', 'real', [],
-                   interface=DataSymbol.Argument(
-                       access=DataSymbol.Access.READ)),
-        DataSymbol)
-    my_mod = ContainerSymbol("my_mod")
-    assert isinstance(
-        DataSymbol('a', 'deferred',
-                   interface=DataSymbol.FortranGlobal(
-                       my_mod, access=DataSymbol.Access.READ)),
-        DataSymbol)
-    dim = DataSymbol('dim', 'integer', [])
-    assert isinstance(DataSymbol('a', 'real', [dim]), DataSymbol)
-    assert isinstance(DataSymbol('a', 'real', [3, dim, None]), DataSymbol)
-
-    # Test with invalid arguments
-    with pytest.raises(NotImplementedError) as error:
-        DataSymbol('a', 'invalidtype', [], 'local')
-    assert (
-        "DataSymbol can only be initialised with {0} datatypes but found "
-        "'invalidtype'.".format(str(DataSymbol.valid_data_types))) in str(
-            error.value)
-
-    with pytest.raises(TypeError) as error:
-        DataSymbol('a', 'real', shape=dim)
-    assert "DataSymbol shape attribute must be a list." in str(error.value)
-
-    with pytest.raises(TypeError) as error:
-        DataSymbol('a', 'real', ['invalidshape'])
-    assert ("DataSymbol shape list elements can only be 'DataSymbol', "
-            "'integer' or 'None'.") in str(error.value)
-
-    with pytest.raises(TypeError) as error:
-        bad_dim = DataSymbol('dim', 'real', [])
-        DataSymbol('a', 'real', [bad_dim], 'local')
-    assert ("Symbols that are part of another symbol shape can "
-            "only be scalar integers, but found") in str(error.value)
-
-    with pytest.raises(TypeError) as error:
-        bad_dim = DataSymbol('dim', 'integer', [3])
-        DataSymbol('a', 'real', [bad_dim], 'local')
-    assert ("Symbols that are part of another symbol shape can "
-            "only be scalar integers, but found") in str(error.value)
-
-    with pytest.raises(ValueError) as error:
-        DataSymbol('a', 'integer', interface=DataSymbol.Argument(),
-                   constant_value=9)
-    assert ("Symbol with a constant value is currently limited to having "
-            "local scope but found 'global'.") in str(error)
-
-    with pytest.raises(ValueError) as error:
-        DataSymbol('a', 'integer', shape=[None], constant_value=9)
-    assert ("Symbol with a constant value must be a scalar but the shape "
-            "attribute is not empty.") in str(error)
-
-    with pytest.raises(ValueError) as error:
-        DataSymbol('a', 'integer', constant_value=9.81)
-    assert ("This DataSymbol instance's datatype is 'integer' which means the "
-            "constant value is expected to be") in str(error)
-    assert "'int'>' but found " in str(error)
-    assert "'float'>'." in str(error)
-
-    with pytest.raises(ValueError) as error:
-        DataSymbol('a', 'character', constant_value=42)
-    assert ("This DataSymbol instance's datatype is 'character' which means "
-            "the constant value is expected to be") in str(error)
-    assert "'str'>' but found " in str(error)
-    assert "'int'>'." in str(error)
-
-    with pytest.raises(ValueError) as error:
-        DataSymbol('a', 'boolean', constant_value="hello")
-    assert ("This DataSymbol instance's datatype is 'boolean' which means the "
-            "constant value is expected to be") in str(error)
-    assert "'bool'>' but found " in str(error)
-    assert "'str'>'." in str(error)
-
-
-def test_symbol_map():
-    '''Test the mapping variable in the DataSymbol class does not raise any
-    exceptions when it is used with the valid_data_types variable in
-    the DataSymbol class.
-
-    '''
-    # "deferred" is not supported in the mapping so we expect
-    # it to have 1 fewer entries than there are valid data types
-    assert len(DataSymbol.valid_data_types) == len(DataSymbol.mapping) + 1
-    for data_type in DataSymbol.valid_data_types:
-        if data_type not in ["deferred"]:
-            assert data_type in DataSymbol.mapping
-
-
-def test_symbol_can_be_printed():
-    '''Test that a DataSymbol instance can always be printed. (i.e. is
-    initialised fully.)'''
-    symbol = DataSymbol("sname", "real")
-    assert "sname: <real, Scalar, local>" in str(symbol)
-
-    sym1 = DataSymbol("s1", "integer")
-    assert "s1: <integer, Scalar, local>" in str(sym1)
-
-    sym2 = DataSymbol("s2", "real", [None, 2, sym1])
-    assert "s2: <real, Array['Unknown bound', 2, s1], local>" in str(sym2)
-
-    my_mod = ContainerSymbol("my_mod")
-    sym3 = DataSymbol("s3", "real",
-                      interface=DataSymbol.FortranGlobal(my_mod))
-    assert ("s3: <real, Scalar, global=FortranModule(my_mod)"
-            in str(sym3))
-
-    sym2._shape.append('invalid')
-    with pytest.raises(InternalError) as error:
-        _ = str(sym2)
-    assert ("DataSymbol shape list elements can only be 'DataSymbol', "
-            "'integer' or 'None', but found") in str(error.value)
-
-    sym3 = DataSymbol("s3", "integer", constant_value=12)
-    assert "s3: <integer, Scalar, local, constant_value=12>" in str(sym3)
-
-
-def test_symbol_constant_value_setter():
-    '''Test that a DataSymbol constant value can be set if given a new valid
-    constant value. Also test that is_constant returns True
-
-    '''
-
-    # Test with valid constant value
-    sym = DataSymbol('a', 'integer', constant_value=7)
-    assert sym.constant_value == 7
-    sym.constant_value = 9
-    assert sym.constant_value == 9
-
-    sym = DataSymbol('a', 'real', constant_value=3.1415)
-    assert sym.constant_value == 3.1415
-    sym.constant_value = 1.0
-    assert sym.constant_value == 1.0
-
-    sym = DataSymbol('a', 'deferred')
-    with pytest.raises(ValueError) as error:
-        sym.constant_value = 1.0
-    assert ("A constant value is not supported for datatype "
-            "'deferred'.") in str(error)
-
-
-def test_symbol_is_constant():
-    '''Test that the DataSymbol is_constant property returns True if a
-    constant value is set and False if it is not.
-
-    '''
-    sym = DataSymbol('a', 'integer')
-    assert not sym.is_constant
-    sym.constant_value = 9
-    assert sym.is_constant
-
-
-def test_symbol_scalar_array():
-    '''Test that the DataSymbol property is_scalar returns True if the
-    DataSymbol is a scalar and False if not and that the DataSymbol property
-    is_array returns True if the DataSymbol is an array and False if not.
-
-    '''
-    sym1 = DataSymbol("s1", "integer")
-    sym2 = DataSymbol("s2", "real", [None, 2, sym1])
-    assert sym1.is_scalar
-    assert not sym1.is_array
-    assert not sym2.is_scalar
-    assert sym2.is_array
-
-
-def test_symbol_invalid_interface():
-    ''' Check that the DataSymbol.interface setter rejects the supplied value
-    if it is not a SymbolInterface. '''
-    sym = DataSymbol("some_var", "real")
-    with pytest.raises(TypeError) as err:
-        sym.interface = "invalid interface spec"
-    assert ("interface to a DataSymbol must be a SymbolInterface or None but"
-            in str(err))
-
-
-def test_symbol_interface():
-    ''' Check the interface getter on a DataSymbol. '''
-    my_mod = ContainerSymbol("my_mod")
-    symbol = DataSymbol("some_var", "real",
-                        interface=DataSymbol.FortranGlobal(my_mod))
-    assert symbol.interface.container_symbol.name == "my_mod"
-
-
-def test_symbol_interface_access():
-    ''' Tests for the SymbolInterface.access setter. '''
-    my_mod = ContainerSymbol("my_mod")
-    symbol = DataSymbol("some_var", "real",
-                        interface=DataSymbol.FortranGlobal(my_mod))
-    symbol.interface.access = DataSymbol.Access.READ
-    assert symbol.interface.access == DataSymbol.Access.READ
-    # Force the error by supplying a string instead of a SymbolAccess type.
-    with pytest.raises(TypeError) as err:
-        symbol.interface.access = "read"
-    assert "must be a 'DataSymbol.Access' but got " in str(err)
-
-
-def test_symbol_argument_str():
-    ''' Check the __str__ method of the DataSymbol.Argument class. '''
-    # A DataSymbol.Argument represents a routine argument by default.
-    interface = DataSymbol.Argument()
-    assert str(interface) == "Argument(pass-by-value=False)"
-
-
-def test_fortranglobal_str():
-    ''' Test the __str__ method of DataSymbol.FortranGlobal. '''
-    # If it's not an argument then we have nothing else to say about it (since
-    # other options are language specific and are implemented in sub-classes).
-    my_mod = ContainerSymbol("my_mod")
-    interface = DataSymbol.FortranGlobal(my_mod)
-    assert str(interface) == "FortranModule(my_mod)"
-
-
-def test_fortranglobal_modname():
-    ''' Test the FortranGlobal.module_name setter error conditions. '''
-    with pytest.raises(TypeError) as err:
-        _ = DataSymbol.FortranGlobal(None)
-    assert ("FortranGlobal container_symbol parameter must be of type"
-            " ContainerSymbol, but found ") in str(err)
-
-
-def test_symbol_copy():
-    '''Test that the DataSymbol copy method produces a faithful separate copy
-    of the original symbol.
-
-    '''
-    symbol = DataSymbol("myname", "real", shape=[1, 2], constant_value=None,
-                        interface=DataSymbol.Argument(
-                            access=DataSymbol.Access.READWRITE))
-    new_symbol = symbol.copy()
-
-    # Check the new symbol has the same properties as the original
-    assert symbol.name == new_symbol.name
-    assert symbol.datatype == new_symbol.datatype
-    assert symbol.shape == new_symbol.shape
-    assert symbol.scope == new_symbol.scope
-    assert symbol.constant_value == new_symbol.constant_value
-    assert symbol.interface == new_symbol.interface
-
-    # Change the properties of the new symbol and check the original
-    # is not affected. Can't check constant_value yet as we have a
-    # shape value
-    new_symbol._name = "new"
-    new_symbol._datatype = "integer"
-    new_symbol.shape[0] = 3
-    new_symbol.shape[1] = 4
-    new_symbol._interface = None
-
-    assert symbol.name == "myname"
-    assert symbol.datatype == "real"
-    assert symbol.shape == [1, 2]
-    assert symbol.scope == "global"
-    assert not symbol.constant_value
-
-    # Now check constant_value
-    new_symbol._shape = []
-    new_symbol.constant_value = True
-
-    assert symbol.shape == [1, 2]
-    assert not symbol.constant_value
-
-
-def test_symbol_copy_properties():
-    '''Test that the DataSymbol copy_properties method works as expected.'''
-
-    symbol = DataSymbol("myname", "real", shape=[1, 2], constant_value=None,
-                        interface=DataSymbol.Argument(
-                            access=DataSymbol.Access.READWRITE))
-
-    # Check an exception is raised if an incorrect argument is passed
-    # in
-    with pytest.raises(TypeError) as excinfo:
-        symbol.copy_properties(None)
-    assert ("Argument should be of type 'DataSymbol' but found 'NoneType'."
-            "") in str(excinfo.value)
-
-    new_symbol = DataSymbol("other_name", "integer", shape=[],
-                            constant_value=7)
-
-    symbol.copy_properties(new_symbol)
-
-    assert symbol.name == "myname"
-    assert symbol.datatype == "integer"
-    assert symbol.shape == []
-    assert symbol.scope == "local"
-    assert symbol.constant_value == 7
-
-
-# Test SymbolTable Class
-
-def test_symboltable_add():
-    '''Test that the add method inserts new symbols in the symbol
-    table, but raises appropiate errors when provided with wrong parameters
-    or duplicate declarations.'''
-    sym_table = SymbolTable()
-
-    # Declare a symbol
-    my_mod = ContainerSymbol("my_mod")
-    sym_table.add(my_mod)
-    sym_table.add(DataSymbol("var1", "real", shape=[5, 1],
-                             interface=DataSymbol.FortranGlobal(
-                                 my_mod,
-                                 access=DataSymbol.Access.READWRITE)))
-    assert sym_table._symbols["my_mod"].name == "my_mod"
-    assert sym_table._symbols["var1"].name == "var1"
-    assert sym_table._symbols["var1"].datatype == "real"
-    assert sym_table._symbols["var1"].shape == [5, 1]
-    assert sym_table._symbols["var1"].scope == "global"
-    assert sym_table._symbols["var1"].access is DataSymbol.Access.READWRITE
-    assert sym_table._symbols["var1"].interface.container_symbol == my_mod
-
-    # Declare a duplicate name symbol
-    with pytest.raises(KeyError) as error:
-        sym_table.add(DataSymbol("var1", "real"))
-    assert ("Symbol table already contains a symbol with name "
-            "'var1'.") in str(error.value)
-
-
-def test_symboltable_swap_symbol_properties():
-    ''' Test the symboltable swap_properties method '''
-
-    symbol1 = DataSymbol("var1", "integer", shape=[], constant_value=7)
-    symbol2 = DataSymbol("dim1", "integer",
-                         interface=DataSymbol.Argument(
-                             access=DataSymbol.Access.READ))
-    symbol3 = DataSymbol("dim2", "integer",
-                         interface=DataSymbol.Argument(
-                             access=DataSymbol.Access.READ))
-    symbol4 = DataSymbol("var2", "real", shape=[symbol2, symbol3],
-                         interface=DataSymbol.Argument(
-                             access=DataSymbol.Access.READWRITE))
-    sym_table = SymbolTable()
-    sym_table.add(symbol1)
-
-    # Raise exception if the first argument is not a symbol
-    with pytest.raises(TypeError) as excinfo:
-        sym_table.swap_symbol_properties(None, symbol1)
-    assert ("Arguments should be of type 'Symbol' but found 'NoneType'."
-            "") in str(excinfo.value)
-
-    # Raise exception if the second argument is not a symbol
-    with pytest.raises(TypeError) as excinfo:
-        sym_table.swap_symbol_properties(symbol1, "symbol")
-    assert ("Arguments should be of type 'Symbol' but found 'str'."
-            "") in str(excinfo.value)
-
-    # Raise exception if the first symbol does not exist in the symbol table
-    with pytest.raises(KeyError) as excinfo:
-        sym_table.swap_symbol_properties(symbol4, symbol1)
-    assert "Symbol 'var2' is not in the symbol table." in str(excinfo.value)
-
-    # Raise exception if the second symbol does not exist in the symbol table
-    with pytest.raises(KeyError) as excinfo:
-        sym_table.swap_symbol_properties(symbol1, symbol4)
-    assert "Symbol 'var2' is not in the symbol table." in str(excinfo.value)
-
-    # Raise exception if both symbols have the same name
-    with pytest.raises(ValueError) as excinfo:
-        sym_table.swap_symbol_properties(symbol1, symbol1)
-    assert("The symbols should have different names, but found 'var1' for "
-           "both.") in str(excinfo.value)
-
-    sym_table.add(symbol2)
-    sym_table.add(symbol3)
-    sym_table.add(symbol4)
-    sym_table.specify_argument_list([symbol2, symbol3, symbol4])
-
-    # Check that properties are swapped
-    sym_table.swap_symbol_properties(symbol1, symbol4)
-
-    assert symbol1.name == "var1"
-    assert symbol1.datatype == "real"
-    assert symbol1.shape == [symbol2, symbol3]
-    assert symbol1.scope == "global"
-    assert symbol1.constant_value is None
-    assert symbol1.interface.access == DataSymbol.Access.READWRITE
-
-    assert symbol4.name == "var2"
-    assert symbol4.datatype == "integer"
-    assert not symbol4.shape
-    assert symbol4.scope == "local"
-    assert symbol4.constant_value == 7
-    assert not symbol4.interface
-
-    # Check symbol references are unaffected
-    sym_table.swap_symbol_properties(symbol2, symbol3)
-    assert symbol1.shape[0].name == "dim1"
-    assert symbol1.shape[1].name == "dim2"
-
-    # Check argument positions are updated. The original positions
-    # were [dim1, dim2, var2]. They should now be [dim2, dim1, var1]
-    assert sym_table.argument_list[0].name == "dim2"
-    assert sym_table.argument_list[1].name == "dim1"
-    assert sym_table.argument_list[2].name == "var1"
-
-
-def test_symboltable_lookup():
-    '''Test that the lookup method retrieves symbols from the symbol table
-    if the name exists, otherwise it raises an error.'''
-    sym_table = SymbolTable()
-    sym_table.add(DataSymbol("var1", "real", shape=[None, None]))
-    sym_table.add(DataSymbol("var2", "integer", shape=[]))
-    sym_table.add(DataSymbol("var3", "real", shape=[]))
-
-    assert isinstance(sym_table.lookup("var1"), DataSymbol)
-    assert sym_table.lookup("var1").name == "var1"
-    assert isinstance(sym_table.lookup("var2"), DataSymbol)
-    assert sym_table.lookup("var2").name == "var2"
-    assert isinstance(sym_table.lookup("var3"), DataSymbol)
-    assert sym_table.lookup("var3").name == "var3"
-
-    with pytest.raises(KeyError) as error:
-        sym_table.lookup("notdeclared")
-    assert "Could not find 'notdeclared' in the Symbol Table." in \
-        str(error.value)
-
-
-def test_symboltable_view(capsys):
-    '''Test the view method of the SymbolTable class, it should print to
-    standard out a representation of the full SymbolTable.'''
-    sym_table = SymbolTable()
-    sym_table.add(DataSymbol("var1", "real"))
-    sym_table.add(DataSymbol("var2", "integer"))
-    sym_table.view()
-    output, _ = capsys.readouterr()
-    assert "Symbol Table:\n" in output
-    assert "var1" in output
-    assert "var2" in output
-
-
-def test_symboltable_can_be_printed():
-    '''Test that a SymbolTable instance can always be printed. (i.e. is
-    initialised fully)'''
-    sym_table = SymbolTable()
-    my_mod = ContainerSymbol("my_mod")
-    sym_table.add(my_mod)
-    sym_table.add(DataSymbol("var1", "real"))
-    sym_table.add(DataSymbol("var2", "integer"))
-    sym_table.add(DataSymbol("var3", "deferred",
-                             interface=DataSymbol.FortranGlobal(my_mod)))
-    sym_table_text = str(sym_table)
-    assert "Symbol Table:\n" in sym_table_text
-    assert "var1" in sym_table_text
-    assert "var2" in sym_table_text
-    assert "\nmy_mod" in sym_table_text
-    assert "FortranModule(my_mod)" in sym_table_text
-
-
-def test_symboltable_specify_argument_list():
-    '''Test that the specify argument list method sets the argument_list
-    with references to each DataSymbol and updates the DataSymbol attributes
-    when needed.'''
-    sym_table = SymbolTable()
-    sym_v1 = DataSymbol("var1", "real", [])
-    sym_table.add(sym_v1)
-    sym_table.add(DataSymbol("var2", "real", []))
-    sym_v1.interface = DataSymbol.Argument(access=DataSymbol.Access.UNKNOWN)
-    sym_table.specify_argument_list([sym_v1])
-
-    assert len(sym_table.argument_list) == 1
-    assert sym_table.argument_list[0].scope == 'global'
-    assert sym_table.argument_list[0].access == DataSymbol.Access.UNKNOWN
-
-    # Test that repeated calls still produce a valid argument list
-    sym_table.specify_argument_list([sym_v1])
-    assert len(sym_table.argument_list) == 1
-
-    # Check that specifying the Interface allows us to specify how
-    # the argument is accessed
-    sym_v2 = sym_table.lookup("var2")
-    sym_v2.interface = DataSymbol.Argument(access=DataSymbol.Access.READWRITE)
-    sym_table.specify_argument_list([sym_v1, sym_v2])
-    assert sym_table.argument_list[1].scope == 'global'
-    assert sym_table.argument_list[1].access == DataSymbol.Access.READWRITE
-
-
-def test_symboltable_specify_argument_list_errors():
-    ''' Check that supplying specify_argument_list() with DataSymbols that
-    don't have the correct Interface information raises the expected
-    errors. '''
-    sym_table = SymbolTable()
-    sym_table.add(DataSymbol("var1", "real", []))
-    sym_table.add(DataSymbol("var2", "real", []))
-    sym_v1 = sym_table.lookup("var1")
-    # Attempt to say the argument list consists of "var1" which at this
-    # point is just a local variable.
-    with pytest.raises(ValueError) as err:
-        sym_table.specify_argument_list([sym_v1])
-    assert "Symbol 'var1:" in str(err)
-    assert ("is listed as a kernel argument but has no associated "
-            "Interface" in str(err))
-    # Now add an Interface for "var1" but of the wrong type
-    sym_v1.interface = DataSymbol.FortranGlobal(ContainerSymbol("my_mod"))
-    with pytest.raises(ValueError) as err:
-        sym_table.specify_argument_list([sym_v1])
-    assert "Symbol 'var1:" in str(err)
-    assert "has an interface of type '" in str(err)
-
-
-def test_symboltable_argument_list_errors():
-    ''' Tests the internal sanity checks of the SymbolTable.argument_list
-    property. '''
-    sym_table = SymbolTable()
-    sym_table.add(DataSymbol("var1", "real", []))
-    sym_table.add(DataSymbol("var2", "real", []))
-    sym_table.add(DataSymbol("var3", "real",
-                             interface=DataSymbol.FortranGlobal(
-                                 ContainerSymbol("my_mod"))))
-    # Manually put a local symbol into the internal list of arguments
-    sym_table._argument_list = [sym_table.lookup("var1")]
-    with pytest.raises(ValueError) as err:
-        sym_table._validate_arg_list(sym_table._argument_list)
-    pattern = ("Symbol \'var1.*\' is listed as a kernel argument but has "
-               "no associated Interface")
-    assert re.search(pattern, str(err)) is not None
-    # Check that the argument_list property converts this error into an
-    # InternalError
-    with pytest.raises(InternalError) as err:
-        _ = sym_table.argument_list
-    assert re.search(pattern, str(err)) is not None
-    # Check that we reject a symbol imported from a module
-    with pytest.raises(ValueError) as err:
-        sym_table._validate_arg_list([sym_table.lookup("var3")])
-    # Manually put that symbol into the argument list
-    sym_table._argument_list = [sym_table.lookup("var3")]
-    pattern = (r"Symbol \'var3.*\' is listed as a kernel argument but has an "
-               r"interface of type \'.*\.FortranGlobal\'>")
-    assert re.search(pattern, str(err)) is not None
-    # Check that the argument_list property converts this error into an
-    # InternalError
-    with pytest.raises(InternalError) as err:
-        _ = sym_table.argument_list
-    assert re.search(pattern, str(err)) is not None
-    # Check that we get the expected TypeError if we provide a list containing
-    # objects that are not Symbols
-    with pytest.raises(TypeError) as err:
-        sym_table._validate_arg_list(["Not a symbol"])
-    assert "Expected a list of DataSymbols but found an object of type" \
-        in str(err)
-
-
-def test_symboltable_validate_non_args():
-    ''' Checks for the validation of non-argument entries in the
-    SymbolTable. '''
-    sym_table = SymbolTable()
-    sym_table.add(DataSymbol("var1", "real", []))
-    sym_table.add(DataSymbol("var2", "real", []))
-    sym_table.add(DataSymbol("var3", "real",
-                             interface=DataSymbol.FortranGlobal(
-                                 ContainerSymbol("my_mod"))))
-    # Everything should be fine so far
-    sym_table._validate_non_args()
-    # Add an entry with an Argument interface
-    sym_table.add(DataSymbol("var4", "real",
-                             interface=DataSymbol.Argument()))
-    # Since this symbol isn't in the argument list, the SymbolTable
-    # is no longer valid
-    with pytest.raises(ValueError) as err:
-        sym_table._validate_non_args()
-    pattern = (r"Symbol 'var4.* is not listed as a kernel argument and yet "
-               "has a DataSymbol.Argument interface")
-    assert re.search(pattern, str(err)) is not None
-
-
-def test_symboltable_contains():
-    '''Test that the __contains__ method returns True if the given name
-    is in the SymbolTable, otherwise returns False.'''
-    sym_table = SymbolTable()
-
-    sym_table.add(DataSymbol("var1", "real", []))
-    sym_table.add(DataSymbol("var2", "real", [None]))
-
-    assert "var1" in sym_table
-    assert "var2" in sym_table
-    assert "var3" not in sym_table
-
-
-def test_symboltable_symbols():
-    '''Test that the symbols property returns a list of the symbols in the
-    SymbolTable.'''
-    sym_table = SymbolTable()
-    assert sym_table.symbols == []
-    sym_table.add(DataSymbol("var1", "real", []))
-    sym_table.add(DataSymbol("var2", "real", [None]))
-    assert len(sym_table.symbols) == 2
-    sym_table.add(DataSymbol("var3", "real", [],
-                             interface=DataSymbol.FortranGlobal(
-                                 ContainerSymbol("my_mod"))))
-    assert len(sym_table.symbols) == 3
-
-
-def test_symboltable_local_variables():
-    '''Test that the local_variables property returns a list with the
-    symbols with local scope.'''
-    sym_table = SymbolTable()
-    assert [] == sym_table.local_variables
-
-    sym_table.add(DataSymbol("var1", "real", []))
-    sym_table.add(DataSymbol("var2", "real", [None]))
-    sym_table.add(DataSymbol("var3", "real", []))
-
-    assert len(sym_table.local_variables) == 3
-    assert sym_table.lookup("var1") in sym_table.local_variables
-    assert sym_table.lookup("var2") in sym_table.local_variables
-    assert sym_table.lookup("var3") in sym_table.local_variables
-    sym_v1 = sym_table.lookup("var1")
-    sym_v1.interface = DataSymbol.Argument(access=DataSymbol.Access.READWRITE)
-    sym_table.specify_argument_list([sym_v1])
-
-    assert len(sym_table.local_variables) == 2
-    assert sym_table.lookup("var1") not in sym_table.local_variables
-    assert sym_table.lookup("var2") in sym_table.local_variables
-    assert sym_table.lookup("var3") in sym_table.local_variables
-
-    sym_table.add(DataSymbol("var4", "real", [],
-                             interface=DataSymbol.FortranGlobal(
-                                 ContainerSymbol("my_mod"))))
-    assert len(sym_table.local_variables) == 2
-    assert sym_table.lookup("var4") not in sym_table.local_variables
-
-
-def test_symboltable_global_variables():
-    ''' Test that the global_variables property returns those DataSymbols with
-    'global' scope (i.e. that represent data that exists outside the current
-    scoping unit) but are not routine arguments. '''
-    sym_table = SymbolTable()
-    assert sym_table.global_variables == []
-    # Add some local symbols
-    sym_table.add(DataSymbol("var1", "real", []))
-    sym_table.add(DataSymbol("var2", "real", [None]))
-    assert sym_table.global_variables == []
-    # Add some global symbols
-    sym_table.add(DataSymbol("gvar1", "real", [],
-                             interface=DataSymbol.FortranGlobal(
-                                 ContainerSymbol("my_mod"))))
-    assert sym_table.lookup("gvar1") in sym_table.global_variables
-    sym_table.add(
-        DataSymbol("gvar2", "real", [],
-                   interface=DataSymbol.Argument(
-                       access=DataSymbol.Access.READWRITE)))
-    gsymbols = sym_table.global_variables
-    assert len(gsymbols) == 1
-    assert sym_table.lookup("gvar2") not in gsymbols
-
-
-def test_symboltable_abstract_properties():
-    '''Test that the SymbolTable abstract properties raise the appropriate
-    error.'''
-    sym_table = SymbolTable()
-
-    with pytest.raises(NotImplementedError) as error:
-        _ = sym_table.data_arguments
-    assert "Abstract property. Which symbols are data arguments is " \
-        "API-specific." in str(error.value)
-
-    with pytest.raises(NotImplementedError) as error:
-        _ = sym_table.iteration_indices
-    assert "Abstract property. Which symbols are iteration indices is " \
-        "API-specific." in str(error.value)
 
 
 def test_modified_kern_line_length(kernel_outputdir, monkeypatch):
